trillium/requirements.txt
chromadb
qdrant-client>=1.7.0
openai>=1.2.0
anthropic>=0.18.0
google-genai>=0.2.0
python-dotenv
pymupdf
pillow
pytesseract
numpy
tqdm
rich
google-cloud-vision
openpyxl
pandas
python-docx
pillow-heif
msal>=1.24.0
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.17.0
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.0
